auth_api/pyproject.toml
[tool.poetry]
name = "auth-sprint-1"
version = "0.1.0"
description = ""
authors = ["ntrafimov <ntrafimov@ozon.ru>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.116.0,<0.117.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
uvicorn = ">=0.35.0,<0.36.0"
pydantic = ">=2.5.3,<3.0.0"
pydantic-settings = ">=2.10.1,<3.0.0"
alembic = ">=1.16.3,<2.0.0"
redis = ">=6.2.0,<7.0.0"
typer = ">=0.16.0,<0.17.0"
passlib = ">=1.7.4,<2.0.0"
python-jose = "==3.3.0"
email-validator = "^2.0"
greenlet = "^3.2.3"
bcrypt = "^4.3.0"
opentelemetry-sdk = ">=1.12,<2.0"
opentelemetry-exporter-otlp = ">=1.12,<2.0"
opentelemetry-instrumentation-fastapi = "0.56b0"
deprecated = "^1.2.14"
orjson = "^3.9.15"
itsdangerous = "^2.2.0"
authlib = "^1.6.1"
aiokafka = "^0.11.0"


[tool.poetry.group.test.dependencies]
pytest = ">=7.0.0,<8.0.0"
pytest-asyncio = ">=0.21.0,<1.0.0"
httpx = ">=0.28.0,<1.0.0"
anyio = ">=4.0.0,<5.0.0"
asyncpg = ">=0.27.0"


[tool.poetry.group.dev.dependencies]
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
